requirements.txt
Flask
flask-cors
phonenumbers
gspread
google-auth
google-cloud-firestore
python-dotenv
